Two more usize that should be u64

+++ src/bam/mod.rs
@@ -819,7 +819,7 @@ impl HeaderView {
 
             let rec = htslib::sam_hdr_parse((l_text + 1) as u64, text as *const i8);
             (*rec).text = text as *mut i8;
-            (*rec).l_text = l_text as usize;
+            (*rec).l_text = l_text as u64;
             rec
         };
